+++ app/src/main/java/com/shub39/grit/core/presentation/settings/ui/SettingsGraph.kt
@@ -18,6 +18,7 @@ import androidx.navigation.compose.rememberNavController
 import com.shub39.grit.core.presentation.settings.SettingsAction
 import com.shub39.grit.core.presentation.settings.SettingsState
 import com.shub39.grit.core.presentation.settings.ui.section.BackupPage
+import com.shub39.grit.core.presentation.settings.ui.section.Changelog
 import com.shub39.grit.core.presentation.settings.ui.section.LookAndFeelPage
 import com.shub39.grit.core.presentation.settings.ui.section.RootPage
 import com.shub39.grit.core.shared_ui.PageFill
@@ -36,8 +37,8 @@ private sealed interface SettingsRoutes {
     @Serializable
     data object Backup: SettingsRoutes
 
-//    @Serializable
-//    data object Server: SettingsRoutes
+    @Serializable
+    data object Changelog: SettingsRoutes
 }
 
 @Composable
@@ -76,8 +77,8 @@ fun SettingsGraph(
                 onAction = onAction,
                 onNavigateToLookAndFeel = { navController.navigate(SettingsRoutes.LookAndFeel) },
                 onNavigateToBackup = { navController.navigate(SettingsRoutes.Backup) },
-                onNavigateToPaywall = onNavigateToPaywall
-//                onNavigateToServer = { navController.navigate(SettingsRoutes.Server) }
+                onNavigateToPaywall = onNavigateToPaywall,
+                onNavigateToChangelog = { navController.navigate(SettingsRoutes.Changelog) }
             )
         }
 
@@ -99,12 +100,12 @@ fun SettingsGraph(
             )
         }
 
-//        composable<SettingsRoutes.Server> {
-//            ServerPage(
-//                state = state,
-//                onNavigateBack = { navController.navigateUp() }
-//            )
-//        }
+        composable<SettingsRoutes.Changelog> {
+            Changelog(
+                changelog = state.changelog,
+                onNavigateBack = { navController.navigateUp() }
+            )
+        }
     }
 }