+++ app/src/main/java/com/shub39/grit/widgets/WidgetSize.kt
@@ -1,0 +1,10 @@
+package com.shub39.grit.widgets
+
+import androidx.compose.ui.unit.dp
+
+data object WidgetSize {
+    val Width2 = 109.dp
+    val Width4 = 245.dp
+    val Height1 = 56.dp
+    val Height2 = 115.dp
+}