+++ app/src/main/java/com/shub39/grit/widgets/habit_overview_widget/HabitOverviewWidgetReceiver.kt
@@ -1,0 +1,8 @@
+package com.shub39.grit.widgets.habit_overview_widget
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+
+class HabitOverviewWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = HabitOverviewWidget()
+}